16: Move action Rules to SettingsController
