+++ NugetAnalyzer.Web/Controllers/HomeController.cs
@@ -1,19 +1,9 @@
-﻿using System;
-using System.Threading.Tasks;
-using Microsoft.AspNetCore.Mvc;
-using Microsoft.Extensions.Options;
-using NugetAnalyzer.BLL.Interfaces;
-using NugetAnalyzer.Common.Configurations;
+﻿using Microsoft.AspNetCore.Mvc;
+
 namespace NugetAnalyzer.Web.Controllers
 {
     public class HomeController : Controller
     {
-        private readonly PackageVersionConfigurations packageVersionConfigurations;
-
-        public HomeController(IOptions<PackageVersionConfigurations> packageVersionConfigurations)
-        {
-            this.packageVersionConfigurations = (packageVersionConfigurations ?? throw new ArgumentNullException(nameof(packageVersionConfigurations))).Value;
-        }
         public IActionResult Index()
         {
             return View();
@@ -27,11 +17,6 @@ namespace NugetAnalyzer.Web.Controllers
         public IActionResult About()
         {
             return View();
-        }
-
-        public IActionResult Rules()
-        {
-            return View(packageVersionConfigurations);
         }
     }
 }
