+++ NugetAnalyzer.Web/wwwroot/stylesheets/Settings/Rules.scss
